requirements.txt
fastapi[all]==0.78.0
pyyaml==6.0